app/requirements.txt
# Chat4All API Hub - Dependencies
# Python 3.11+ required

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Message Broker
kafka-python==2.0.2

# Object Storage
minio==7.2.0

# Security
bcrypt==4.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
python-dotenv==1.0.0
